refactor foundry setup into modules
