+++ scripts/setup-foundry/setup_foundry/__init__.py
@@ -1,0 +1,1 @@
+"""Verified Foundry setup tooling."""
