+++ scripts/setup-foundry/setup_foundry/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Foundry setup subcommands."""
